requirement.txt
groq
streamlit
duckduckgo-search

pypdf
sentence-transformers
faiss-cpu